backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
redis>=5.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.3.0
aiohttp>=3.9.0
httpx>=0.26.0
email-validator>=2.1.0
bcrypt>=4.1.0
pyyaml>=6.0.1
cryptography>=42.0.0